Fixed SQS Failure in tests
Apparently moving to 1.44 to 1.51 for the AWS SDK causes
a decoding error of some sorts.

+++ pkg/pubsub/pubsub_suite_test.go
@@ -77,8 +77,12 @@ var _ = BeforeSuite(func() {
 }, 2.0)
 
 var _ = AfterSuite(func() {
-	Ω(awsLocal.Terminate(context.Background())).ToNot(HaveOccurred())
-	Ω(redisContainer.Terminate(context.Background())).ToNot(HaveOccurred())
+	if awsLocal != nil {
+		Ω(awsLocal.Terminate(context.Background())).ToNot(HaveOccurred())
+	}
+	if redisContainer != nil {
+		Ω(redisContainer.Terminate(context.Background())).ToNot(HaveOccurred())
+	}
 }, 2.0)
 
 // getQueueName provides a way to obtain a process-independent name for the SQS queue,
